move some CL deps to trampoline
